libnsgif: Don't use ssize_t (it's POSIX, and not in C99)

+++ libnsgif/libnsgif.c
@@ -123,8 +123,8 @@ static gif_result
 gif_initialise_frame_extensions(gif_animation *gif, const int frame)
 {
         const unsigned char *gif_data, *gif_end;
-        ssize_t gif_bytes;
-        ssize_t block_size;
+        intmax_t gif_bytes;
+        intmax_t block_size;
 
         /* Get our buffer position etc.	*/
         gif_data = (const unsigned char *)(gif->gif_data + gif->buffer_position);
@@ -255,10 +255,10 @@ static gif_result gif_initialise_frame(gif_animation *gif)
         gif_frame *temp_buf;
 
         const unsigned char *gif_data, *gif_end;
-        ssize_t gif_bytes;
+        intmax_t gif_bytes;
         unsigned int flags = 0;
         unsigned int width, height, offset_x, offset_y;
-        ssize_t block_size, colour_table_size;
+        intmax_t block_size, colour_table_size;
         bool first_image = true;
         gif_result return_value;
         bool premature_eof = false;
@@ -437,7 +437,7 @@ static gif_result gif_initialise_frame(gif_animation *gif)
                 if (gif_bytes < 1) return GIF_INSUFFICIENT_FRAME_DATA;
                 block_size = gif_data[0] + 1;
                 /* Check if the frame data runs off the end of the file	*/
-                if ((ssize_t)(gif_bytes - block_size) < 0) {
+                if ((intmax_t)(gif_bytes - block_size) < 0) {
                         /* Try to recover by signaling the end of the gif.
                          * Once we get garbage data, there is no logical way to
                          * determine where the next frame is.  It's probably
@@ -486,8 +486,8 @@ static gif_result gif_initialise_frame(gif_animation *gif)
 static gif_result gif_skip_frame_extensions(gif_animation *gif)
 {
         const unsigned char *gif_data, *gif_end;
-        ssize_t gif_bytes;
-        ssize_t block_size;
+        intmax_t gif_bytes;
+        intmax_t block_size;
 
         /* Get our buffer position etc.	*/
         gif_data = (const unsigned char *)(gif->gif_data + gif->buffer_position);
@@ -587,13 +587,13 @@ gif_internal_decode_frame(gif_animation *gif,
 {
         unsigned int index = 0;
         const unsigned char *gif_data, *gif_end;
-        ssize_t gif_bytes;
+        intmax_t gif_bytes;
         unsigned int width, height, offset_x, offset_y;
         unsigned int flags, colour_table_size, interlace;
         unsigned int *colour_table;
         unsigned int *frame_data = 0;	// Set to 0 for no warnings
         unsigned int *frame_scanline;
-        ssize_t save_buffer_position;
+        intmax_t save_buffer_position;
         unsigned int return_value = 0;
         unsigned int x, y, decode_y, burst_bytes;
         register unsigned char colour;
